Network.AWS.STS.DecodeAuthorizationMessage

Description

Decodes additional information about the authorization status of a request from an encoded message returned in response to an AWS request.

For example, if a user is not authorized to perform an action that he or she has requested, the request returns a UnauthorizedOperation response (an HTTP 403 response). Some AWS actions additionally return an encoded message that can provide details about this authorization failure.

The message is encoded because the details of the authorization status can constitute privileged information that the user who requested the action should not see. To decode an authorization status message, a user must be granted permissions via an IAM policy to request the DecodeAuthorizationMessage ('sts:DecodeAuthorizationMessage') action.

The decoded message includes the following type of information:

Whether the request was denied due to an explicit deny or due to the absence of an explicit allow. For more information, see Determining Whether aRequest is Allowed or Denied in Using IAM. The principal who made the request. The requested action. The requested resource. The values of condition keys in the context of the user's request.

http://docs.aws.amazon.com/STS/latest/APIReference/API_DecodeAuthorizationMessage.html

Request

data DecodeAuthorizationMessage

Request constructor

decodeAuthorizationMessage

Arguments

:: Text
damEncodedMessage

-> DecodeAuthorizationMessage

DecodeAuthorizationMessage constructor.

The fields accessible through corresponding lenses are:

damEncodedMessage :: Text

Request lenses

damEncodedMessage :: Lens' DecodeAuthorizationMessage Text

The encoded message that was returned with the response.

Response

data DecodeAuthorizationMessageResponse

Response constructor

decodeAuthorizationMessageResponse :: DecodeAuthorizationMessageResponse

DecodeAuthorizationMessageResponse constructor.

The fields accessible through corresponding lenses are:

damrDecodedMessage :: Maybe Text

Response lenses

damrDecodedMessage :: Lens' DecodeAuthorizationMessageResponse (Maybe Text)

An XML document that contains the decoded message. For more information, see DecodeAuthorizationMessage.