Network.AWS.STS.GetCallerIdentity

Description

Returns details about the IAM identity whose credentials are used to call the API.

Creating a Request

getCallerIdentity :: GetCallerIdentity

Creates a value of GetCallerIdentity with the minimum fields required to make a request.

data GetCallerIdentity

See: getCallerIdentity smart constructor.

Destructuring the Response

getCallerIdentityResponse

Arguments

:: Int
gcirsResponseStatus

-> GetCallerIdentityResponse

Creates a value of GetCallerIdentityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcirsARN - The AWS ARN associated with the calling entity.

gcirsAccount - The AWS account ID number of the account that owns or contains the calling entity.

gcirsUserId - The unique identifier of the calling entity. The exact value depends on the type of entity making the call. The values returned are those listed in the aws:userid column in the Principal table found on the Policy Variables reference page in the IAM User Guide .

gcirsResponseStatus - -- | The response status code.

data GetCallerIdentityResponse

Contains the response to a successful GetCallerIdentity request, including information about the entity making the request.

See: getCallerIdentityResponse smart constructor.

Response Lenses

gcirsARN :: Lens' GetCallerIdentityResponse (Maybe Text)

The AWS ARN associated with the calling entity.

gcirsAccount :: Lens' GetCallerIdentityResponse (Maybe Text)

The AWS account ID number of the account that owns or contains the calling entity.

gcirsUserId :: Lens' GetCallerIdentityResponse (Maybe Text)

The unique identifier of the calling entity. The exact value depends on the type of entity making the call. The values returned are those listed in the aws:userid column in the Principal table found on the Policy Variables reference page in the IAM User Guide .

gcirsResponseStatus :: Lens' GetCallerIdentityResponse Int

- | The response status code.