Network.AWS.STS.Types

Description

The AWS Security Token Service (STS) is a web service that enables you to request temporary, limited-privilege credentials for AWS Identity and Access Management (IAM) users or for users that you authenticate (federated users).

Service

data STS

Supported version (2011-06-15) of the AWS Security Token Service service.

Errors

data STSError

A sum type representing possible errors returned by the STS service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

ExpiredTokenException
The web identity token that was passed is expired or is not valid. Get a new identity token from the identity provider and then retry the request.

Fields

_eteMessage :: Maybe Text

IDPCommunicationErrorException
The request could not be fulfilled because the non-AWS identity provider (IDP) that was asked to verify the incoming identity token could not be reached. This is often a transient error caused by network conditions. Retry the request a limited number of times so that you don't exceed the request rate. If the error persists, the non-AWS identity provider might be down or not responding.

Fields

_idpceeMessage :: Maybe Text

IDPRejectedClaimException
The identity provider (IdP) reported that authentication failed. This might be because the claim is invalid. If this error is returned for the AssumeRoleWithWebIdentity operation, it can also mean that the claim has expired or has been explicitly revoked.

Fields

_idprceMessage :: Maybe Text

InvalidAuthorizationMessageException
The error returned if the message passed to DecodeAuthorizationMessage was invalid. This can happen if the token contains invalid characters, such as linebreaks.

Fields

_iameMessage :: Maybe Text

InvalidIdentityTokenException
The web identity token that was passed could not be validated by AWS. Get a new identity token from the identity provider and then retry the request.

Fields

_iiteMessage :: Maybe Text

MalformedPolicyDocumentException
The request was rejected because the policy document was malformed. The error message describes the specific error.

Fields

_mpdeMessage :: Maybe Text

PackedPolicyTooLargeException
The request was rejected because the policy document was too large. The error message describes how big the policy document is, in packed form, as a percentage of what the API allows.

Fields

_pptleMessage :: Maybe Text

STSClient HttpException

STSSerializer String

STSService String

_ExpiredTokenException :: Prism' STSError (Maybe Text)

The web identity token that was passed is expired or is not valid. Get a new identity token from the identity provider and then retry the request.

See: ExpiredTokenException

_IDPCommunicationErrorException :: Prism' STSError (Maybe Text)

The request could not be fulfilled because the non-AWS identity provider (IDP) that was asked to verify the incoming identity token could not be reached. This is often a transient error caused by network conditions. Retry the request a limited number of times so that you don't exceed the request rate. If the error persists, the non-AWS identity provider might be down or not responding.

See: IDPCommunicationErrorException

_IDPRejectedClaimException :: Prism' STSError (Maybe Text)

The identity provider (IdP) reported that authentication failed. This might be because the claim is invalid. If this error is returned for the AssumeRoleWithWebIdentity operation, it can also mean that the claim has expired or has been explicitly revoked.

See: IDPRejectedClaimException

_InvalidAuthorizationMessageException :: Prism' STSError (Maybe Text)

The error returned if the message passed to DecodeAuthorizationMessage was invalid. This can happen if the token contains invalid characters, such as linebreaks.

See: InvalidAuthorizationMessageException

_InvalidIdentityTokenException :: Prism' STSError (Maybe Text)

The web identity token that was passed could not be validated by AWS. Get a new identity token from the identity provider and then retry the request.

See: InvalidIdentityTokenException

_MalformedPolicyDocumentException :: Prism' STSError (Maybe Text)

The request was rejected because the policy document was malformed. The error message describes the specific error.

See: MalformedPolicyDocumentException

_PackedPolicyTooLargeException :: Prism' STSError (Maybe Text)

The request was rejected because the policy document was too large. The error message describes how big the policy document is, in packed form, as a percentage of what the API allows.

See: PackedPolicyTooLargeException

_STSClient :: Prism' STSError HttpException

See: STSClient

_STSSerializer :: Prism' STSError String

See: STSSerializer

_STSService :: Prism' STSError String

See: STSService

XML

xmlOptions :: Tagged a XMLOptions

AssumedRoleUser

data AssumedRoleUser

The Amazon Resource Name (ARN) and the assumed role ID, which are identifiers that you can use to refer to the resulting temporary security credentials. For example, you can reference these credentials as a principal in a resource-based policy by using the ARN or assumed role ID. The ARN and ID include the RoleSessionName that you specified when you called AssumeRole.

assumedRoleUser

Arguments

:: Text
aruAssumedRoleId

-> Text
aruArn

-> AssumedRoleUser

Smart constructor for the minimum required fields to construct a valid AssumedRoleUser data type.

AssumedRoleUser is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AssumedRoleId :: Text

Arn :: Text

aruAssumedRoleId :: Lens' AssumedRoleUser Text

A unique identifier that contains the role ID and the role session name of the role that is being assumed. The role ID is generated by AWS when the role is created.

aruArn :: Lens' AssumedRoleUser Text

The ARN of the temporary security credentials that are returned from the AssumeRole action. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.

Credentials

data Credentials

The session credentials for API authentication.

credentials

Arguments

:: Text
cAccessKeyId

-> Text
cSecretAccessKey

-> Text
cSessionToken

-> ISO8601
cExpiration

-> Credentials

Smart constructor for the minimum required fields to construct a valid Credentials data type.

Credentials is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AccessKeyId :: Text

SecretAccessKey :: Text

SessionToken :: Text

Expiration :: ISO8601

cAccessKeyId :: Lens' Credentials Text

The access key ID that identifies the temporary security credentials.

cSecretAccessKey :: Lens' Credentials Text

The secret access key that can be used to sign requests.

cSessionToken :: Lens' Credentials Text

The token that users must pass to the service API to use the temporary credentials.

cExpiration :: Lens' Credentials ISO8601

The date on which the current credentials expire.

FederatedUser

data FederatedUser

Identifiers for the federated user associated with the credentials (such as arn:aws:sts::123456789012:federated-user/Bob or 123456789012:Bob). You can use the federated user's ARN in your resource-based policies, such as an Amazon S3 bucket policy.

federatedUser

Arguments

:: Text
fuFederatedUserId

-> Text
fuArn

-> FederatedUser

Smart constructor for the minimum required fields to construct a valid FederatedUser data type.

FederatedUser is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

FederatedUserId :: Text

Arn :: Text

fuFederatedUserId :: Lens' FederatedUser Text

The string that identifies the federated user associated with the credentials, similar to the unique ID of an IAM user.

fuArn :: Lens' FederatedUser Text

The ARN that specifies the federated user that is associated with the credentials. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.