Network.AWS.STS.DecodeAuthorizationMessage

Description

Decodes additional information about the authorization status of a request from an encoded message returned in response to an AWS request.

For example, if a user is not authorized to perform an action that he or she has requested, the request returns a Client.UnauthorizedOperation response (an HTTP 403 response). Some AWS actions additionally return an encoded message that can provide details about this authorization failure.

The message is encoded because the details of the authorization status can constitute privileged information that the user who requested the action should not see. To decode an authorization status message, a user must be granted permissions via an IAM policy to request the DecodeAuthorizationMessage (sts:DecodeAuthorizationMessage ) action.

The decoded message includes the following type of information:

Whether the request was denied due to an explicit deny or due to the absence of an explicit allow. For more information, see Determining Whether a Request is Allowed or Denied in the IAM User Guide .

The principal who made the request.

The requested action.

The requested resource.

The values of condition keys in the context of the user's request.

Creating a Request

decodeAuthorizationMessage

Arguments

:: Text
damEncodedMessage

-> DecodeAuthorizationMessage

Creates a value of DecodeAuthorizationMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

damEncodedMessage - The encoded message that was returned with the response.

data DecodeAuthorizationMessage

See: decodeAuthorizationMessage smart constructor.

Request Lenses

damEncodedMessage :: Lens' DecodeAuthorizationMessage Text

The encoded message that was returned with the response.

Destructuring the Response

decodeAuthorizationMessageResponse

Arguments

:: Int
damrsResponseStatus

-> DecodeAuthorizationMessageResponse

Creates a value of DecodeAuthorizationMessageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

damrsDecodedMessage - An XML document that contains the decoded message.

damrsResponseStatus - -- | The response status code.

data DecodeAuthorizationMessageResponse

A document that contains additional information about the authorization status of a request from an encoded message that is returned in response to an AWS request.

See: decodeAuthorizationMessageResponse smart constructor.

Response Lenses

damrsDecodedMessage :: Lens' DecodeAuthorizationMessageResponse (Maybe Text)

An XML document that contains the decoded message.

damrsResponseStatus :: Lens' DecodeAuthorizationMessageResponse Int

- | The response status code.