Network.AWS.STS.Types

Service

data STS

Version 2011-06-15 of the Amazon Security Token Service service.

Error

data RESTError :: *

XML

ns :: Text

Credentials

data Credentials

credentials

Arguments

:: Text
cAccessKeyId

-> Text
cSecretAccessKey

-> Text
cSessionToken

-> UTCTime
cExpiration

-> Credentials

Credentials constructor.

The fields accessible through corresponding lenses are:

cAccessKeyId :: Text

cExpiration :: UTCTime

cSecretAccessKey :: Text

cSessionToken :: Text

cAccessKeyId :: Lens' Credentials Text

The access key ID that identifies the temporary security credentials.

cExpiration :: Lens' Credentials UTCTime

The date on which the current credentials expire.

cSecretAccessKey :: Lens' Credentials Text

The secret access key that can be used to sign requests.

cSessionToken :: Lens' Credentials Text

The token that users must pass to the service API to use the temporary credentials.

FederatedUser

data FederatedUser

federatedUser

Arguments

:: Text
fuFederatedUserId

-> Text
fuArn

-> FederatedUser

FederatedUser constructor.

The fields accessible through corresponding lenses are:

fuArn :: Text

fuFederatedUserId :: Text

fuArn :: Lens' FederatedUser Text

The ARN that specifies the federated user that is associated with the credentials. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.

fuFederatedUserId :: Lens' FederatedUser Text

The string that identifies the federated user associated with the credentials, similar to the unique ID of an IAM user.

AssumedRoleUser

data AssumedRoleUser

assumedRoleUser

Arguments

:: Text
aruAssumedRoleId

-> Text
aruArn

-> AssumedRoleUser

AssumedRoleUser constructor.

The fields accessible through corresponding lenses are:

aruArn :: Text

aruAssumedRoleId :: Text

aruArn :: Lens' AssumedRoleUser Text

The ARN of the temporary security credentials that are returned from the AssumeRole action. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.

aruAssumedRoleId :: Lens' AssumedRoleUser Text

A unique identifier that contains the role ID and the role session name of the role that is being assumed. The role ID is generated by AWS when the role is created.