Network.AWS.STS.V2011_06_15.Types

Description

The AWS Security Token Service (STS) is a web service that enables you to request temporary, limited-privilege credentials for AWS Identity and Access Management (IAM) users or for users that you authenticate (federated users).

Documentation

data STS

Supported version (2011-06-15) of the AWS Security Token Service service.

xmlOptions :: Tagged a XMLOptions

data AssumedRoleUser

The Amazon Resource Name (ARN) and the assumed role ID, which are identifiers that you can use to refer to the resulting temporary security credentials. For example, you can reference these credentials as a principal in a resource-based policy by using the ARN or assumed role ID. The ARN and ID include the RoleSessionName that you specified when you called AssumeRole.

Constructors

AssumedRoleUser

Fields

_aruArn :: Text

The ARN of the temporary security credentials that are returned from the AssumeRole action. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.

_aruAssumedRoleId :: Text

A unique identifier that contains the role ID and the role session name of the role that is being assumed. The role ID is generated by AWS when the role is created.

data Credentials

The temporary security credentials, which include an access key ID, a secret access key, and a security (or session) token.

Constructors

Credentials

Fields

_cSecretAccessKey :: Text

The secret access key that can be used to sign requests.

_cSessionToken :: Text

The token that users must pass to the service API to use the temporary credentials.

_cExpiration :: ISO8601

The date on which the current credentials expire.

_cAccessKeyId :: Text

The access key ID that identifies the temporary security credentials.

data FederatedUser

Identifiers for the federated user associated with the credentials (such as arn:aws:sts::123456789012:federated-user/Bob or 123456789012:Bob). You can use the federated user's ARN in your resource-based policies, such as an Amazon S3 bucket policy.

Constructors

FederatedUser

Fields

_fuArn :: Text

The ARN that specifies the federated user that is associated with the credentials. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.

_fuFederatedUserId :: Text

The string that identifies the federated user associated with the credentials, similar to the unique ID of an IAM user.

aruAssumedRoleId :: Lens' AssumedRoleUser Text

aruArn :: Lens' AssumedRoleUser Text

cSessionToken :: Lens' Credentials Text

cSecretAccessKey :: Lens' Credentials Text

cExpiration :: Lens' Credentials ISO8601

cAccessKeyId :: Lens' Credentials Text

fuFederatedUserId :: Lens' FederatedUser Text

fuArn :: Lens' FederatedUser Text