Network.AWS.STS.V2011_06_15.Types

Description

The AWS Security Token Service (STS) is a web service that enables you to request temporary, limited-privilege credentials for AWS Identity and Access Management (IAM) users or for users that you authenticate (federated users).

Service

data STS

Supported version (2011-06-15) of the AWS Security Token Service service.

XML

xmlOptions :: Tagged a XMLOptions

AssumedRoleUser

data AssumedRoleUser

The Amazon Resource Name (ARN) and the assumed role ID, which are identifiers that you can use to refer to the resulting temporary security credentials. For example, you can reference these credentials as a principal in a resource-based policy by using the ARN or assumed role ID. The ARN and ID include the RoleSessionName that you specified when you called AssumeRole.

mkAssumedRoleUser

Arguments

:: Text
aruAssumedRoleId

-> Text
aruArn

-> AssumedRoleUser

Smart constructor for the minimum required fields to construct a valid AssumedRoleUser data type.

AssumedRoleUser is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AssumedRoleId :: Text

Arn :: Text

aruAssumedRoleId :: Lens' AssumedRoleUser Text

A unique identifier that contains the role ID and the role session name of the role that is being assumed. The role ID is generated by AWS when the role is created.

aruArn :: Lens' AssumedRoleUser Text

The ARN of the temporary security credentials that are returned from the AssumeRole action. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.

Credentials

data Credentials

The temporary security credentials, which include an access key ID, a secret access key, and a security (or session) token.

mkCredentials

Arguments

:: Text
cAccessKeyId

-> Text
cSecretAccessKey

-> Text
cSessionToken

-> ISO8601
cExpiration

-> Credentials

Smart constructor for the minimum required fields to construct a valid Credentials data type.

Credentials is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AccessKeyId :: Text

SecretAccessKey :: Text

SessionToken :: Text

Expiration :: ISO8601

cAccessKeyId :: Lens' Credentials Text

The access key ID that identifies the temporary security credentials.

cSecretAccessKey :: Lens' Credentials Text

The secret access key that can be used to sign requests.

cSessionToken :: Lens' Credentials Text

The token that users must pass to the service API to use the temporary credentials.

cExpiration :: Lens' Credentials ISO8601

The date on which the current credentials expire.

FederatedUser

data FederatedUser

Identifiers for the federated user associated with the credentials (such as arn:aws:sts::123456789012:federated-user/Bob or 123456789012:Bob). You can use the federated user's ARN in your resource-based policies, such as an Amazon S3 bucket policy.

mkFederatedUser

Arguments

:: Text
fuFederatedUserId

-> Text
fuArn

-> FederatedUser

Smart constructor for the minimum required fields to construct a valid FederatedUser data type.

FederatedUser is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

FederatedUserId :: Text

Arn :: Text

fuFederatedUserId :: Lens' FederatedUser Text

The string that identifies the federated user associated with the credentials, similar to the unique ID of an IAM user.

fuArn :: Lens' FederatedUser Text

The ARN that specifies the federated user that is associated with the credentials. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in Using IAM.